+++ x-pack/legacy/plugins/xpack_main/server/lib/xpack_info_license.test.js
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { licensingMock } from '../../../../../plugins/licensing/server/licensing.mock';
+import { licensingMock } from '../../../../../plugins/licensing/server/licensing.mocks';
 import { XPackInfoLicense } from './xpack_info_license';
 
 function getXPackInfoLicense(getRawLicense) {
