+++ x-pack/plugins/licensing/common/licensing.mocks.ts
